+++ packages/tab-scroller/src/mwc-tab-scroller-base.ts
@@ -15,9 +15,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 import {matches} from '@material/dom/ponyfill';
-import {addHasRemoveClass, BaseElement, eventOptions, html, query} from '@material/mwc-base/base-element';
+import {addHasRemoveClass, BaseElement} from '@material/mwc-base/base-element';
 import {MDCTabScrollerAdapter} from '@material/tab-scroller/adapter';
 import MDCTabScrollerFoundation from '@material/tab-scroller/foundation.js';
+import {eventOptions, html, query} from 'lit-element';
 
 export class TabScrollerBase extends BaseElement {
   protected mdcFoundation!: MDCTabScrollerFoundation;